gimbal_bridge refactored to pass all ament_lint tests
